feat: [SCRUM-154] folder and import refactor
